+++ infra/utils/utils.go
@@ -5,12 +5,12 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func GetUserID(c *gin.Context) (int64, error) {
+func GetUserID(c *gin.Context) (float64, error) {
 	v, exists := c.Get("user_id")
 	if !exists {
 		return 0, errors.New("unauthorized")
 	}
-	id, ok := v.(int64)
+	id, ok := v.(float64)
 	if !ok {
 		return 0, errors.New("invalid user_id")
 	}
